Rework to make more flexible
