+++ src/main/java/com/patreon/models/campaign/misc/simple/SimpleReward.java
@@ -1,0 +1,9 @@
+package com.patreon.models.campaign.misc.simple;
+
+import com.github.jasminb.jsonapi.annotations.Meta;
+import com.patreon.models.common.LinkedModel;
+
+public class SimpleReward extends LinkedModel {
+    @Meta
+    private SimpleRewards.SimpleRewardData data;
+}
