+++ src/main/java/com/patreon/models/common/BaseIdModel.java
@@ -1,0 +1,12 @@
+package com.patreon.models.common;
+
+import com.github.jasminb.jsonapi.annotations.Id;
+
+public class BaseIdModel {
+    @Id
+    private String id;
+
+    public String getId() {
+        return id;
+    }
+}
