+++ src/main/java/com/patreon/models/user/UserIdObject.java
@@ -1,0 +1,9 @@
+package com.patreon.models.user;
+
+public class UserIdObject {
+    private String user_id;
+
+    public String getUserId() {
+        return user_id;
+    }
+}
